// frontend/src/api/generated/model/productType.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * EstacionCafe API
 * API REST de EstacionCafé. Todas las rutas requieren JWT (header `Authorization: Bearer <token>` o cookie `auth_token`) excepto login y logout. Los roles permitidos de cada operación están en su descripción y en `x-roles`.
 * OpenAPI spec version: 2.0.0
 */

export interface ProductType {
  productTypeId: number;
  name: string;
}